FAIL TO RUN TASK
What I trying to do here is to have a screen after the choice on dots
direction is made, that instructs subject to press A for reporting
a CP, and B for reporting no CP. And catch that report.

I tried to use the topsTaskHelperMessage class to display that instruction
screen with an entry function in the state machine. Since I did not
want to clear the screen after displaying the instructions, I added the
method shownoclear to the topsTaskHelperMessage class, but this generates
a bug (see console log from 29 Oct 11:29 (rouhgly)