Revert to task version that doesn't store the dots
the file modularTasks/tasks/topsTreeNodeTaskReversingDots4AFC.m in the
parent commit was set-up to store the dots positions from each frame on
every trial in a dotsInfo task property. This setup works with the
dotsDrawableDotKinetogramDebug class for the dots stimulus.

In the present commit, the file modularTasks/tasks/topsTreeNodeTaskReversingDots4AFC.m
should be used with the dotsDrawableDotKinetogram class, and the dots
positions on every frame will NOT be recorded. However, the trialData
struct (and therefore the FIRA matrix) will contain the number of frames
drawn before the CP (numberDrawPreCP) and the number of frames drawn after
CP time (numberDrawPostCP).